refactor and fix test
