+++ app/src/androidTest/java/com/github/houseorganizer/houseorganizer/EditHouseholdActivityTest.java
@@ -10,9 +10,9 @@ import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
 import static androidx.test.espresso.matcher.ViewMatchers.isClickable;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.isEnabled;
-import static androidx.test.espresso.matcher.ViewMatchers.withEffectiveVisibility;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
+import static com.github.houseorganizer.houseorganizer.CustomMatchers.withBitmap;
 import static com.github.houseorganizer.houseorganizer.FirebaseTestsHelper.TEST_HOUSEHOLD_NAMES;
 import static com.github.houseorganizer.houseorganizer.FirebaseTestsHelper.TEST_USERS_EMAILS;
 import static com.github.houseorganizer.houseorganizer.FirebaseTestsHelper.UNKNOWN_USER;
@@ -21,34 +21,21 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import android.content.Context;
 import android.content.Intent;
-import android.content.res.Resources;
 import android.graphics.Bitmap;
-import android.graphics.Canvas;
-import android.graphics.Point;
-import android.graphics.drawable.Drawable;
-import android.view.View;
-import android.widget.ImageView;
 
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.espresso.intent.Intents;
-import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
-import androidx.test.platform.app.InstrumentationRegistry;
 
-import com.github.houseorganizer.houseorganizer.house.EditHousehold;
+import com.github.houseorganizer.houseorganizer.house.EditHouseholdActivity;
 import com.github.houseorganizer.houseorganizer.house.HouseSelectionActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.zxing.WriterException;
 
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -61,7 +48,7 @@ import java.util.Map;
 import java.util.concurrent.ExecutionException;
 
 @RunWith(AndroidJUnit4.class)
-public class EditHouseholdTest {
+public class EditHouseholdActivityTest {
     private static FirebaseFirestore db;
     private static FirebaseAuth auth;
     private static Intent intentFromHouseSelection;
@@ -74,7 +61,7 @@ public class EditHouseholdTest {
 
         db = FirebaseFirestore.getInstance();
         auth = FirebaseAuth.getInstance();
-        intentFromHouseSelection = new Intent(ApplicationProvider.getApplicationContext(), EditHousehold.class).putExtra(HouseSelectionActivity.HOUSEHOLD_TO_EDIT, TEST_HOUSEHOLD_NAMES[0]);
+        intentFromHouseSelection = new Intent(ApplicationProvider.getApplicationContext(), EditHouseholdActivity.class).putExtra(HouseSelectionActivity.HOUSEHOLD_TO_EDIT, TEST_HOUSEHOLD_NAMES[0]);
     }
 
     @AfterClass
@@ -84,7 +71,7 @@ public class EditHouseholdTest {
     }
 
     @Rule
-    public ActivityScenarioRule<EditHousehold> editHouseholdRule = new ActivityScenarioRule<>(intentFromHouseSelection);
+    public ActivityScenarioRule<EditHouseholdActivity> editHouseholdRule = new ActivityScenarioRule<>(intentFromHouseSelection);
 
     @Before
     public void setupHouseholds() throws ExecutionException, InterruptedException {
@@ -158,43 +145,9 @@ public class EditHouseholdTest {
     public void showQRCodeDisplaysCorrectQRCode() throws WriterException {
         // Perform clicks
         onView(withId(R.id.showQRCode)).perform(click());
-        Context cx = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        Point size = new Point();
-        ((AppCompatActivity)cx).getWindowManager().getDefaultDisplay().getSize(size);
-        final int padding = 50;
-        int length = size.x - padding;
-        Bitmap expected = EditHousehold.createQRCodeBitmap(TEST_HOUSEHOLD_NAMES[0], length);
+        int length = 2000;
+        Bitmap expected = EditHouseholdActivity.createQRCodeBitmap(TEST_HOUSEHOLD_NAMES[0], length);
         onView(withId(R.id.image_dialog)).check(matches(withBitmap(expected)));
-    }
-
-    public static Matcher<View> withBitmap(final Bitmap expected) {
-        return new TypeSafeMatcher<View>() {
-            @Override
-            protected boolean matchesSafely(View target) {
-                if (!(target instanceof ImageView)){
-                    return false;
-                }
-                ImageView imageView = (ImageView) target;
-
-                Bitmap bitmap = getBitmap (imageView.getDrawable());
-
-                return bitmap.sameAs(expected);
-            }
-
-            private Bitmap getBitmap(Drawable drawable){
-                Bitmap bitmap = Bitmap.createBitmap(drawable.getIntrinsicWidth(),
-                        drawable.getIntrinsicHeight(), Bitmap.Config.ARGB_8888);
-                Canvas canvas = new Canvas(bitmap);
-                drawable.setBounds(0, 0, canvas.getWidth(), canvas.getHeight());
-                drawable.draw(canvas);
-                return bitmap;
-            }
-
-            @Override
-            public void describeTo(Description description) {
-                description.appendText("ImageView with bitmap same as bitmap " + expected.toString());
-            }
-        };
     }
 
     @Test
