+++ app/src/androidTest/java/com/github/houseorganizer/houseorganizer/LoginEmailTest.java
@@ -29,8 +29,6 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.util.concurrent.ExecutionException;
-
 
 @RunWith(AndroidJUnit4.class)
 public class LoginEmailTest {
