+++ app/src/main/AndroidManifest.xml
@@ -67,7 +67,7 @@
             android:exported="false"
             android:parentActivityName=".house.HouseSelectionActivity" />
         <activity
-            android:name=".house.EditHousehold"
+            android:name=".house.EditHouseholdActivity"
             android:exported="false"
             android:parentActivityName=".house.HouseSelectionActivity" />
