+++ app/src/main/java/com/github/houseorganizer/houseorganizer/NavBar/NavBarHelpers.java
@@ -4,8 +4,8 @@ import android.content.Context;
 import android.content.Intent;
 
 import com.github.houseorganizer.houseorganizer.panels.CalendarActivity;
-import com.github.houseorganizer.houseorganizer.panels.MainScreenActivity;
 import com.github.houseorganizer.houseorganizer.panels.GroceriesActivity;
+import com.github.houseorganizer.houseorganizer.panels.MainScreenActivity;
 import com.google.firebase.firestore.DocumentReference;
 
 public class NavBarHelpers {
