+++ app/src/main/java/com/github/houseorganizer/houseorganizer/panels/CalendarActivity.java
@@ -17,7 +17,6 @@ import com.google.firebase.firestore.FirebaseFirestore;
 
 import java.time.YearMonth;
 import java.time.format.DateTimeFormatter;
-
 import java.util.OptionalInt;
 
 public final class CalendarActivity extends NavBarActivity {
